requirements-web.txt
fastapi
uvicorn[standard]
jinja2
aiofiles
python-multipart
openai>=0.27.0
numpy
httpx
faiss-cpu
sounddevice
pydub
webrtcvad
pvporcupine==1.9.1
PyYAML
coloredlogs
